@@ -24,7 +24,7 @@ classifiers =
     License :: OSI Approved :: Apache Software License
 
 [options]
-python_requires = >=3.8
+python_requires = >=3.11
 py_modules = envoy.base.utils
 packages = find_namespace:
 install_requires =
