@@ -1,16 +1,16 @@
 
 import pathlib
-from typing import Dict, List, Optional, Tuple, TypedDict
+from typing import TypedDict
 
 from packaging import version as _version
 
-from envoy.base.utils import interface
+from envoy.base.utils import interface  # noqa: F401
 
 
 class Change(str):
 
-    def __add__(self, other) -> "Change":
-        return Change(str(self) + str(other))
+    def __add__(self, other: str) -> "Change":
+        return Change(str(self) + other)
 
 
 class SourceChangeDict(TypedDict):
@@ -23,8 +23,8 @@ class ChangeDict(TypedDict):
     change: Change
 
 
-ChangeList = List[ChangeDict]
-SourceChangeList = List[SourceChangeDict]
+ChangeList = list[ChangeDict]
+SourceChangeList = list[SourceChangeDict]
 
 
 class BaseChangelogDict(TypedDict):
@@ -33,33 +33,33 @@ class BaseChangelogDict(TypedDict):
 
 class ChangelogSourceDict(BaseChangelogDict, total=False):
     # This should match envoy:changelogs/sections.yaml
-    changes: Optional[SourceChangeList]
-    behavior_changes: Optional[SourceChangeList]
-    minor_behavior_changes: Optional[SourceChangeList]
-    bug_fixes: Optional[SourceChangeList]
-    removed_config_or_runtime: Optional[SourceChangeList]
-    new_features: Optional[SourceChangeList]
-    deprecated: Optional[SourceChangeList]
+    changes: SourceChangeList | None
+    behavior_changes: SourceChangeList | None
+    minor_behavior_changes: SourceChangeList | None
+    bug_fixes: SourceChangeList | None
+    removed_config_or_runtime: SourceChangeList | None
+    new_features: SourceChangeList | None
+    deprecated: SourceChangeList | None
 
 
 class ChangelogChangeSectionsDict(TypedDict, total=False):
     # This should match envoy:changelogs/sections.yaml
-    changes: Optional[ChangeList]
-    behavior_changes: Optional[ChangeList]
-    minor_behavior_changes: Optional[ChangeList]
-    bug_fixes: Optional[ChangeList]
-    removed_config_or_runtime: Optional[ChangeList]
-    new_features: Optional[ChangeList]
-    deprecated: Optional[ChangeList]
+    changes: ChangeList | None
+    behavior_changes: ChangeList | None
+    minor_behavior_changes: ChangeList | None
+    bug_fixes: ChangeList | None
+    removed_config_or_runtime: ChangeList | None
+    new_features: ChangeList | None
+    deprecated: ChangeList | None
 
 
 class ChangelogDict(BaseChangelogDict, ChangelogChangeSectionsDict):
     pass
 
 
-ChangelogPathsDict = Dict[_version.Version, pathlib.Path]
-ChangelogsDict = Dict[_version.Version, "interface.IChangelog"]
-MinorVersionsDict = Dict[_version.Version, Tuple[_version.Version, ...]]
+ChangelogPathsDict = dict[_version.Version, pathlib.Path]
+ChangelogsDict = dict[_version.Version, "interface.IChangelog"]
+MinorVersionsDict = dict[_version.Version, tuple[_version.Version, ...]]
 
 
 class BaseChangelogSectionDict(TypedDict):
@@ -70,13 +70,13 @@ class ChangelogSectionDict(BaseChangelogSectionDict, total=False):
     description: str
 
 
-ChangelogSectionsDict = Dict[str, ChangelogSectionDict]
+ChangelogSectionsDict = dict[str, ChangelogSectionDict]
 
-VersionConfigDict = Dict[str, str]
+VersionConfigDict = dict[str, str]
 
 
-ProjectDevResultDict = Dict
-SyncResultDict = Dict[_version.Version, bool]
+ProjectDevResultDict = dict
+SyncResultDict = dict[_version.Version, bool]
 
 
 class ProjectReleaseResultDict(TypedDict):
@@ -110,5 +110,5 @@ class ProjectChangeDict(TypedDict, total=False):
     trigger: ProjectTriggerResultDict
 
 
-VersionDict = Dict[_version.Version, _version.Version]
-InventoryDict = Dict[_version.Version, pathlib.Path]
+VersionDict = dict[_version.Version, _version.Version]
+InventoryDict = dict[_version.Version, pathlib.Path]
